requirements.txt
abstract==2022.7.10
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.3
base32hex==1.0.2
billiard==4.1.0
celery==5.3.4
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorclass==2.2.2
colouration==2020.7.26
cron-descriptor==1.4.0
Django==4.2.6
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-redis==5.4.0
django-timezone-field==6.0.1
docopt==0.6.2
graphviz==0.20.1
idna==3.4
kombu==5.3.2
packaging==23.2
Pillow==10.0.1
pip-upgrader==1.4.15
prompt-toolkit==3.0.39
python-crontab==3.0.0
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
redis==5.0.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
terminaltables==3.1.10
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.6
vine==5.0.0
wcwidth==0.2.8
